requirements.txt
Office365_REST_Python_Client==2.6.2
pdf2image==1.17.0
pdfplumber==0.11.9
pikepdf==10.5.1
pytesseract==0.3.13
